+++ CI/SCRIPTS/test_display.py
@@ -46,7 +46,7 @@ def test_display():
 
     # Test
     np.testing.assert_array_equal(scaled_2d, arr_2d[0, ...])
-    np.testing.assert_array_equal(scaled_3d, arr_3d)
+    np.testing.assert_almost_equal(scaled_3d, arr_3d)
     np.testing.assert_almost_equal(scaled_3d_m, arr_3d_m)  # Almost here
 
     # Scale to uint8